backend/requirements.txt
regex==2021.4.4
googleapiclient
requests==2.25.1
pandas==1.2.4
nltk==3.6.2
Flask==1.1.1
beautifulsoup4==4.9.3
google_api_python_client==2.7.0
protobuf==3.17.1
scikit_learn==0.24.2
google-cloud
google-api-core
bs4~=0.0.1
sklearn~=0.0
scikit-learn~=0.24.2
six~=1.16.0
httplib2~=0.19.1
uritemplate~=3.0.1